Changed servo slider feedback to be more dynamic
Instead of hardcoded checks for the sensor names, use a regex on the structure of valid servo names.
Instead of giving feedback bubbles an ID and then setting their value, search for existing feedback bubbles by the current servo ID.
Instead of a hardcoded condition to not draw the feedback bar for specific slider, use a class for all sliders that mustn't have a feedback bar.